+++ CHANGELOG.md
@@ -39,6 +39,9 @@ to remain consistent with CLI flags. You should specify `code_challenge_method` 
 
 - [#1788](https://github.com/oauth2-proxy/oauth2-proxy/pull/1788) Update base docker image to alpine 3.16
 
+- [#1760](https://github.com/oauth2-proxy/oauth2-proxy/pull/1760) Option to configure API routes
+
+
 # V7.3.0
 
 ## Release Highlights